move Orderbook struct under PoolPairState & to new source file
