+++ blockchain/pdex/state_v2.go
@@ -1,15 +1,11 @@
 package pdex
 
 import (
-	"encoding/json"
 	"errors"
-	"fmt"
-	"math/big"
 	"reflect"
 	"sort"
 	"strconv"
 
-	v2 "github.com/incognitochain/incognito-chain/blockchain/pdex/v2utils"
 	"github.com/incognitochain/incognito-chain/common"
 	"github.com/incognitochain/incognito-chain/dataaccessobject/rawdbv2"
 	"github.com/incognitochain/incognito-chain/dataaccessobject/statedb"
@@ -23,89 +19,10 @@ type stateV2 struct {
 	waitingContributions        map[string]rawdbv2.Pdexv3Contribution
 	deletedWaitingContributions map[string]rawdbv2.Pdexv3Contribution
 	poolPairs                   map[string]PoolPairState //
-	orderbooks                  map[string]Orderbook
 	params                      Params
 	stakingPoolsState           map[string]StakingPoolState // tokenID -> StakingPoolState
 	producer                    stateProducerV2
 	processor                   stateProcessorV2
-}
-
-type Order = rawdbv2.Pdexv3Order
-
-type Orderbook struct {
-	orders []*Order
-}
-
-func (ob Orderbook) MarshalJSON() ([]byte, error) {
-	temp := struct {
-		Orders []*Order `json:"orders"`
-	}{ob.orders}
-	return json.Marshal(temp)
-}
-
-func (ob *Orderbook) UnmarshalJSON(data []byte) error {
-	var temp struct {
-		Orders []*Order `json:"orders"`
-	}
-	err := json.Unmarshal(data, &temp)
-	if err != nil {
-		return err
-	}
-	ob.orders = temp.Orders
-	return nil
-}
-
-// InsertOrder() appends a new order while keeping the list sorted (ascending by Token1Rate / Token0Rate)
-func (ob *Orderbook) InsertOrder(ord *Order) {
-	insertAt := func(lst []*Order, i int, newItem *Order) []*Order {
-		if i == len(lst) {
-			return append(lst, newItem)
-		}
-		lst = append(lst[:i+1], lst[i:]...)
-		lst[i] = newItem
-		return lst
-	}
-	index := sort.Search(len(ob.orders), func(i int) bool {
-		ordRate := big.NewInt(0).SetUint64(ob.orders[i].Token0Rate())
-		ordRate.Mul(ordRate, big.NewInt(0).SetUint64(ord.Token1Rate()))
-		myRate := big.NewInt(0).SetUint64(ob.orders[i].Token1Rate())
-		myRate.Mul(myRate, big.NewInt(0).SetUint64(ord.Token0Rate()))
-		// compare Token1Rate / Token0Rate of current order in the list to ord
-		if ord.TradeDirection() == v2.TradeDirectionSell0 {
-			// orders selling token0 are iterated from start of list (buy the least token1), so we resolve equality of rate by putting the new one last
-			return ordRate.Cmp(myRate) < 0
-		} else {
-			// orders selling token1 are iterated from end of list (buy the least token0), so we resolve equality of rate by putting the new one first
-			return ordRate.Cmp(myRate) <= 0
-		}
-	})
-	ob.orders = insertAt(ob.orders, index, ord)
-}
-
-// NextOrder() returns the matchable order with the best rate that has any outstanding balance to sell
-func (ob *Orderbook) NextOrder(tradeDirection byte) (*v2.MatchingOrder, string, error) {
-	lstLen := len(ob.orders)
-	switch tradeDirection {
-	case v2.TradeDirectionSell0:
-		for i := 0; i < lstLen; i++ {
-			// only match a trade with an order of the opposite direction
-			if ob.orders[i].TradeDirection() != tradeDirection && ob.orders[i].Token1Balance() > 0 {
-				return &v2.MatchingOrder{ob.orders[i]}, ob.orders[i].Id(), nil
-			}
-		}
-		// no active order
-		return nil, "", nil
-	case v2.TradeDirectionSell1:
-		for i := lstLen - 1; i >= 0; i-- {
-			if ob.orders[i].TradeDirection() != tradeDirection && ob.orders[i].Token0Balance() > 0 {
-				return &v2.MatchingOrder{ob.orders[i]}, ob.orders[i].Id(), nil
-			}
-		}
-		// no active order
-		return nil, "", nil
-	default:
-		return nil, "", fmt.Errorf("Invalid trade direction %d", tradeDirection)
-	}
 }
 
 type Params struct {
@@ -146,14 +63,12 @@ func newStateV2WithValue(
 	poolPairs map[string]PoolPairState,
 	params Params,
 	stakingPoolsState map[string]StakingPoolState,
-	orderbooks map[string]Orderbook,
 ) *stateV2 {
 	return &stateV2{
 		waitingContributions:        waitingContributions,
 		deletedWaitingContributions: deletedWaitingContributions,
 		poolPairs:                   poolPairs,
 		stakingPoolsState:           stakingPoolsState,
-		orderbooks:                  orderbooks,
 		params:                      params,
 	}
 }
@@ -203,8 +118,10 @@ func initStateV2(
 			}
 			shares[k] = *NewShareWithValue(value.Amount(), tradingFees, value.LastUpdatedBeaconHeight())
 		}
+		// TODO: read order book from storage
+		orderbook := Orderbook{}
 		poolPair := NewPoolPairStateWithValue(
-			v.Value(), shares,
+			v.Value(), shares, orderbook,
 		)
 		poolPairs[k] = *poolPair
 	}
@@ -214,7 +131,7 @@ func initStateV2(
 		make(map[string]rawdbv2.Pdexv3Contribution),
 		poolPairs,
 		params,
-		nil, nil,
+		nil,
 	), nil
 }
 
@@ -288,8 +205,8 @@ func (s *stateV2) Process(env StateEnvironment) error {
 				s.deletedWaitingContributions,
 			)
 		case metadataCommon.Pdexv3TradeRequestMeta:
-			s.poolPairs, s.orderbooks, err = s.processor.trade(env.StateDB(), inst,
-				s.poolPairs, s.orderbooks,
+			s.poolPairs, err = s.processor.trade(env.StateDB(), inst,
+				s.poolPairs,
 			)
 		default:
 			Logger.log.Debug("Can not process this metadata")
@@ -363,10 +280,9 @@ func (s *stateV2) BuildInstructions(env StateEnvironment) ([][]string, error) {
 	instructions = append(instructions, modifyParamsInstructions...)
 
 	var tradeInstructions [][]string
-	tradeInstructions, s.poolPairs, s.orderbooks, err = s.producer.trade(
+	tradeInstructions, s.poolPairs, err = s.producer.trade(
 		tradeTxs,
 		s.poolPairs,
-		s.orderbooks,
 	)
 	if err != nil {
 		return instructions, err
